moved GetValue to component
