+++ frontend/src/components/GetValue.css
@@ -1,0 +1,6 @@
+
+.GetValue {
+    & p.error-message {
+        color: red;
+    }
+}
